Backend/usuarios/serializers.py
# users/serializers.py
from rest_framework import serializers
from usuarios.models import Usuario

class UsuarioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Usuario
        fields = ('username','nombre','correo', )