Backend/requirements.txt
Django==2.2.4
psycopg2==2.8.3
django-rest-auth==0.9.5
djangorestframework==3.10.2
